finalizando release 1.0.1

See merge request aula-daves/uniacademia/2023-1/back-end/rafael/back-end-vendas!20

+++ src/main/java/br/com/backEndVendas/model/ItemPedido.java
@@ -16,7 +16,7 @@ public class ItemPedido{
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "id_pedido")
     private Pedido pedido;
-    private long idProduto;
+    private int idProduto;
     private int quantidade;
